+++ src/Features/Advertisement/Doctors/repository.js
@@ -96,21 +96,17 @@ const filterAdvertisement = async (query) => {
     let connection = await pool.getConnection();
 
     try {
-        // Define the rangeCondition object based on the query parameters
         const minPrice = query.minPrice ? parseInt(query.minPrice) : undefined;
         const maxPrice = query.maxPrice ? parseInt(query.maxPrice) : undefined;
         const rangeCondition = minPrice !== undefined && maxPrice !== undefined ? { price: { min: minPrice, max: maxPrice } } : {};
 
-        // Remove minPrice and maxPrice from query object
         if (query?.minPrice) delete query.minPrice;
         if (query?.maxPrice) delete query.maxPrice;
 
-        // Call filterQuery with the correct rangeCondition
         const [sql, values] = await filterQuery("doctors", [], { is_active: 1, ...query }, rangeCondition);
         const [rows, fields] = await connection.query(sql, values);
         const data = await parseImages(rows);
-        return data;
-
+        return { error: false, message: "doctors filter list", "data": data };
     } catch (error) {
         logger.info(error);
         throw new ApplicationError(error, 500);
@@ -120,15 +116,15 @@ const filterAdvertisement = async (query) => {
 };
 
 
-export const updateAdvertisement = async (advertisementID, filter) => {
+export const updateAdvertisement = async (advertisementID, updateBody, userId) => {
     let connection = await pool.getConnection();
 
     try {
 
-        if (!filter || typeof filter !== 'object') {
-            throw new ApplicationError("Invalid filter object provided", 400);
+        if (!updateBody || typeof updateBody !== 'object') {
+            throw new ApplicationError("Invalid updateBody object provided", 400);
         }
-        const [sql, values] = await updateQuery("doctors", filter, { "id": advertisementID })
+        const [sql, values] = await updateQuery("doctors", updateBody, { "id": advertisementID, "user_id": userId })
         const [rows, field] = await connection.query(sql, values)
 
         if (!rows) {
@@ -145,16 +141,21 @@ export const updateAdvertisement = async (advertisementID, filter) => {
     }
 };
 
-export const deactivateAdvertisement = async (advertisementID) => {
+export const deactivateAdvertisement = async (advertisementID, userId) => {
     let connection = await pool.getConnection();
-
     try {
-        const sql = `UPDATE doctors SET is_active = 0 WHERE id = ?`
-        const [rows, fields] = await connection.query(sql, [advertisementID]);
-        if (rows.affectedRows === 0) {
-            throw new ApplicationError("doctors not deactivated. No matching doctors found for the provided ID.", 404);
+        const select = `SELECT * FROM doctors WHERE is_active = 1 AND id = ? AND user_id = ?`;
+        const advertisement = await connection.query(select, [advertisementID, userId]);
+
+        // Check if advertisement exists
+        if (!advertisement.length) {
+            throw new ApplicationError("Advertisement not found", 500);
         }
-        return { error: false, message: "doctors deactivated successfully", advertisements: rows };
+
+        const sql = `UPDATE doctors SET is_active = 0 WHERE id = ?`;
+        const [rows, fields] = await connection.query(sql, [advertisementID]);
+
+        return { error: false, message: "Advertisement deactivated successfully" };
     } catch (error) {
         logger.info(error);
         throw new ApplicationError(error, 500);
@@ -163,11 +164,11 @@ export const deactivateAdvertisement = async (advertisementID) => {
     }
 };
 
-export const addImage = async (advertisementID, files) => {
+export const addImage = async (advertisementID, files, userId) => {
     let connection = await pool.getConnection();
 
     try {
-        const [query, values] = await selectQuery("doctors", {}, { id: advertisementID })
+        const [query, values] = await selectQuery("doctors", {}, { id: advertisementID, user_id: userId })
         const [advertisement, field] = await connection.query(query, values);
         if (advertisement.length == 0) {
             throw new ApplicationError("doctors not found.", 404);
@@ -189,12 +190,12 @@ export const addImage = async (advertisementID, files) => {
     }
 };
 
-export const deleteImage = async (advertisementID, files) => {
+export const deleteImage = async (advertisementID, files, userId) => {
     let connection = await pool.getConnection();
 
     try {
-        const sql = `SELECT * FROM doctors WHERE id = ?`
-        const [rows, fields] = await connection.query(sql, [advertisementID])
+        const sql = `SELECT * FROM doctors WHERE id = ? AND user_id = ?`
+        const [rows, fields] = await connection.query(sql, [advertisementID, userId])
 
         if (rows[0].length == 0) {
             throw new ApplicationError("doctors not found.", 404);
@@ -229,12 +230,12 @@ export const deleteImage = async (advertisementID, files) => {
 
 
 
-export const activateAdvertisement = async (advertisementID) => {
+export const activateAdvertisement = async (advertisementID, userId) => {
     let connection = await pool.getConnection();
 
     try {
 
-        const [query, values] = await selectQuery('doctors', { is_active: 1 }, { id: advertisementID })
+        const [query, values] = await selectQuery('doctors', { is_active: 1 }, { id: advertisementID, user_id: userId })
         const [advertisement] = await connection.query(query, values);
 
         if (advertisement.length == 0) {
@@ -251,11 +252,11 @@ export const activateAdvertisement = async (advertisementID) => {
     }
 };
 
-export const deleteAdvertisement = async (advertisementID) => {
+export const deleteAdvertisement = async (advertisementID, userId) => {
     let connection = await pool.getConnection();
     try {
-        const sql = `DELETE FROM doctors WHERE id = ?`
-        const [advertisement] = await connection.query(sql, advertisementID);
+        const sql = `DELETE FROM doctors WHERE id = ? AND user_id = ?`
+        const [advertisement] = await connection.query(sql, [advertisementID, userId]);
         return { error: false, message: "property deleted successfully" };
     } catch (error) {
         logger.info(error);
