+++ src/Features/Advertisement/Hospitality/repository.js
@@ -91,25 +91,22 @@ const getListAdvertisement = async () => {
     }
 };
 
+
 const filterAdvertisement = async (query) => {
     let connection = await pool.getConnection();
 
     try {
-        // Define the rangeCondition object based on the query parameters
         const minPrice = query.minPrice ? parseInt(query.minPrice) : undefined;
         const maxPrice = query.maxPrice ? parseInt(query.maxPrice) : undefined;
         const rangeCondition = minPrice !== undefined && maxPrice !== undefined ? { price: { min: minPrice, max: maxPrice } } : {};
 
-        // Remove minPrice and maxPrice from query object
         if (query?.minPrice) delete query.minPrice;
         if (query?.maxPrice) delete query.maxPrice;
 
-        // Call filterQuery with the correct rangeCondition
         const [sql, values] = await filterQuery("hospitality", [], { is_active: 1, ...query }, rangeCondition);
         const [rows, fields] = await connection.query(sql, values);
         const data = await parseImages(rows);
-        return data;
-
+        return { error: false, message: "hospitality filter list", "data": data };
     } catch (error) {
         logger.info(error);
         throw new ApplicationError(error, 500);
@@ -119,15 +116,15 @@ const filterAdvertisement = async (query) => {
 };
 
 
-export const updateAdvertisement = async (advertisementID, body) => {
+export const updateAdvertisement = async (advertisementID, updateBody, userId) => {
     let connection = await pool.getConnection();
 
     try {
 
-        if (!body || typeof body !== 'object' || body == {}) {
-            throw new ApplicationError("Invalid filter object provided", 400);
+        if (!updateBody || typeof updateBody !== 'object') {
+            throw new ApplicationError("Invalid updateBody object provided", 400);
         }
-        const [sql, values] = await updateQuery("hospitality", body, { "id": advertisementID })
+        const [sql, values] = await updateQuery("hospitality", updateBody, { "id": advertisementID, "user_id": userId })
         const [rows, field] = await connection.query(sql, values)
 
         if (!rows) {
@@ -143,16 +140,21 @@ export const updateAdvertisement = async (advertisementID, body) => {
     }
 };
 
-export const deactivateAdvertisement = async (advertisementID) => {
+export const deactivateAdvertisement = async (advertisementID, userId) => {
     let connection = await pool.getConnection();
-
     try {
-        const sql = `UPDATE hospitality SET is_active = 0 WHERE id = ?`
-        const [rows, fields] = await connection.query(sql, [advertisementID]);
-        if (rows.affectedRows === 0) {
-            throw new ApplicationError("hospitality not deactivated. No matching hospitality found for the provided ID.", 404);
+        const select = `SELECT * FROM hospitality WHERE is_active = 1 AND id = ? AND user_id = ?`;
+        const advertisement = await connection.query(select, [advertisementID, userId]);
+
+        // Check if advertisement exists
+        if (!advertisement.length) {
+            throw new ApplicationError("Advertisement not found", 500);
         }
-        return { error: false, message: "hospitality deactivated successfully", advertisements: rows };
+
+        const sql = `UPDATE hospitality SET is_active = 0 WHERE id = ?`;
+        const [rows, fields] = await connection.query(sql, [advertisementID]);
+
+        return { error: false, message: "Advertisement deactivated successfully" };
     } catch (error) {
         logger.info(error);
         throw new ApplicationError(error, 500);
@@ -161,11 +163,11 @@ export const deactivateAdvertisement = async (advertisementID) => {
     }
 };
 
-export const addImage = async (advertisementID, files) => {
+export const addImage = async (advertisementID, files, userId) => {
     let connection = await pool.getConnection();
 
     try {
-        const [query, values] = await selectQuery("hospitality", {}, { id: advertisementID })
+        const [query, values] = await selectQuery("hospitality", {}, { id: advertisementID, user_id: userId })
         const [advertisement, field] = await connection.query(query, values);
         if (advertisement.length == 0) {
             throw new ApplicationError("hospitality not found.", 404);
@@ -187,12 +189,12 @@ export const addImage = async (advertisementID, files) => {
     }
 };
 
-export const deleteImage = async (advertisementID, files) => {
+export const deleteImage = async (advertisementID, files, userId) => {
     let connection = await pool.getConnection();
 
     try {
-        const sql = `SELECT * FROM hospitality WHERE id = ?`
-        const [rows, fields] = await connection.query(sql, [advertisementID])
+        const sql = `SELECT * FROM hospitality WHERE id = ? AND user_id = ?`
+        const [rows, fields] = await connection.query(sql, [advertisementID, userId])
 
         if (rows[0].length == 0) {
             throw new ApplicationError("hospitality not found.", 404);
@@ -224,26 +226,12 @@ export const deleteImage = async (advertisementID, files) => {
     }
 };
 
-export const listUserAdvertisement = async (userID) => {
+
+
+export const activateAdvertisement = async (advertisementID, userId) => {
     let connection = await pool.getConnection();
-
     try {
-        const advertisements = await connection.query('hospitality').where('user_id', userID);
-        return { error: false, message: "User advertisement list", advertisements };
-    } catch (error) {
-        logger.info(error);
-        throw new ApplicationError(error, 500);
-    } finally {
-        connection.release();
-    }
-};
-
-export const activateAdvertisement = async (advertisementID) => {
-    let connection = await pool.getConnection();
-
-    try {
-
-        const [query, values] = await selectQuery('hospitality', { is_active: 1 }, { id: advertisementID })
+        const [query, values] = await selectQuery('hospitality', { is_active: 1 }, { id: advertisementID, user_id: userId })
         const [advertisement] = await connection.query(query, values);
 
         if (advertisement.length == 0) {
@@ -260,13 +248,12 @@ export const activateAdvertisement = async (advertisementID) => {
     }
 };
 
-export const deleteAdvertisement = async (advertisementID) => {
+export const deleteAdvertisement = async (advertisementID, userId) => {
     let connection = await pool.getConnection();
-
     try {
-        const sql = `DELETE FROM hospitality WHERE id = ?`
-        const [advertisement] = await connection.query(sql, advertisementID);
-        return { error: false, message: "property deleted successfully" };
+        const sql = `DELETE FROM hospitality WHERE id = ? AND user_id = ?`
+        const [advertisement] = await connection.query(sql, [advertisementID, userId]);
+        return { error: false, message: "hospitality deleted successfully" };
     } catch (error) {
         logger.info(error);
         throw new ApplicationError(error, 500);
@@ -274,6 +261,7 @@ export const deleteAdvertisement = async (advertisementID) => {
         connection.release();
     }
 };
+
 
 export default {
     addAdvertisement,
